+++ tests/Unit/Client/ResponseModelAdapter/FileTest.php
@@ -242,6 +242,13 @@ class FileTest extends \PHPUnit\Framework\TestCase
                                     'default' => '600px',
                                     'description' => 'Height of embedded window (px or %)',
                                 ],
+                                'enablestoreurl' => [
+                                    'type' => 'boolean',
+                                    'default' => false,
+                                    'description' => '',
+                                    'paid' => true,
+                                    'reqPlanLevel' => 30,
+                                ]
                             ],
                         ],
                     ],
@@ -453,6 +460,34 @@ class FileTest extends \PHPUnit\Framework\TestCase
                                         'default' => '600px',
                                         'description' => 'Height of embedded window (px or %)',
                                     ],
+                                    'enablestoreurl' => [
+                                        'type' => 'boolean',
+                                        'default' => false,
+                                        'description' => '',
+                                        'paid' => true,
+                                        'reqPlanLevel' => 30,
+                                    ],
+                                    'storeurl' => [
+                                        'type' => 'string',
+                                        'default' => '',
+                                        'description' => '',
+                                        'paid' => true,
+                                        'reqPlanLevel' => 30
+                                    ],
+                                    'hidehints' => [
+                                        'type' => 'boolean',
+                                        'default' => false,
+                                        'description' => 'Hide tutorial hints',
+                                        'paid' => true,
+                                        'reqPlanLevel' => 30
+                                    ],
+                                    'somenewparam' => [
+                                        'type' => 'boolean',
+                                        'default' => false,
+                                        'description' => '',
+                                        'paid' => true,
+                                        'reqPlanLevel' => 30,
+                                    ]
                                 ],
                             ],
                             'bucket' => 'cdn.cappasity3d.com',